% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_groups.R
\name{age_groups}
\alias{age_groups}
\title{age_groups}
\usage{
age_groups(data, column)
}
\arguments{
\item{data}{df}

\item{column}{age column}
}
\value{
column with age groups
}
\description{
age_groups
}
\examples{
mtcars |> age_groups(mpg)
}
