% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_value_label.R
\name{apply_value_label}
\alias{apply_value_label}
\title{apply_value_label}
\usage{
apply_value_label(df, variable, survey_labels, label_column = "value_label")
}
\arguments{
\item{df}{dataframe}

\item{variable}{column to convert}

\item{survey_labels}{list of labels for variable}

\item{label_column}{current labels}
}
\value{
dataframe
}
\description{
apply_value_label
}
\examples{
df |> apply_value_label(col1, label_lst)
}
