% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{as.col_spec}
\alias{as.col_spec}
\title{Coerce to a column specification}
\usage{
as.col_spec(x)
}
\arguments{
\item{x}{Input object}
}
\description{
This is most useful for generating a specification using the short form or coercing from a list.
}
\examples{
as.col_spec("cccnnn")
}
\keyword{internal}
