% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cohen_d.R
\name{calculate_cohen_d}
\alias{calculate_cohen_d}
\title{calculate_cohen_d}
\usage{
calculate_cohen_d(data, grouping_column, y_column)
}
\arguments{
\item{data}{dataframe}

\item{grouping_column}{column to group}

\item{y_column}{y column}
}
\value{
Cohen's D
}
\description{
calculate_cohen_d
}
\examples{
calculate_cohen_d(mtcars, vs, hp)
}
