% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variance.R
\name{check_variance}
\alias{check_variance}
\title{check_variance}
\usage{
check_variance(formula, data)
}
\arguments{
\item{formula}{t-test formula}

\item{data}{dataframe}
}
\value{
variances for different tests
}
\description{
check_variance
}
\examples{
t.test(extra ~ group, data = sleep) |> check_variance()
}
