% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{combine_data}
\usage{
combine_data(df, label_var, img_lab_var, model_var)
}
\arguments{
\item{df}{dataframe}

\item{label_var}{label column}

\item{img_lab_var}{variable relating to the image}

\item{model_var}{model variable}
}
\value{
merged dataframe
}
\description{
combine_data
}
\examples{
df |> combine_data(label_column, img_label_column, model_column)
}
