% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_v.R
\name{cram_v}
\alias{cram_v}
\title{cram_v}
\usage{
cram_v(test)
}
\arguments{
\item{test}{chi square test}
}
\value{
cramer's V
}
\description{
cram_v
}
\examples{
chisq.test(c(A = 20, B = 15, C = 25)) |> cram_v()
}
