% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_variables.R
\name{create_dummy_variables}
\alias{create_dummy_variables}
\title{create_dummy_variables}
\usage{
create_dummy_variables(data, column, prefix = "dummy")
}
\arguments{
\item{data}{dataframe}

\item{column}{column to create dummy variables with}

\item{prefix}{not used}
}
\value{
dataframe with dummy variables included
}
\description{
create_dummy_variables
}
\examples{
create_dummy_variables(mtcars, vs)
}
