% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_labeller.R
\name{custom_labeller}
\alias{custom_labeller}
\title{custom_labeller}
\usage{
custom_labeller(data, facet_var, custom_labels)
}
\arguments{
\item{data}{dataframe}

\item{facet_var}{variable to facet}

\item{custom_labels}{list of labels to use}
}
\value{
labeller object for ggplot2
}
\description{
custom_labeller
}
\examples{
custom_labeller(df, col1, lab_list)
}
