% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_means.R
\name{diff_means}
\alias{diff_means}
\title{diff_means}
\usage{
diff_means(data, condition, value)
}
\arguments{
\item{data}{dataframe}

\item{condition}{condition category}

\item{value}{value for the category}
}
\value{
dataframe of the difference in means
}
\description{
diff_means
}
\examples{
mtcars |> diff_means(vs, hp)
}
