% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_means2.R
\name{diff_means2}
\alias{diff_means2}
\title{diff_means2}
\usage{
diff_means2(data, condition, value)
}
\arguments{
\item{data}{dataframe}

\item{condition}{condition category}

\item{value}{value for the category}
}
\value{
dataframe of the difference in means
}
\description{
diff_means2
}
\examples{
mtcars |> diff_means2(vs, hp)
}
