% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_names2.R
\name{get_col_names2}
\alias{get_col_names2}
\title{get_col_names2}
\usage{
get_col_names2(df, start_col, end_col)
}
\arguments{
\item{df}{dataframe}

\item{start_col}{starting column}

\item{end_col}{last column}
}
\value{
column names between a range
}
\description{
get_col_names2
}
\examples{
get_col_names2(mtcars, cyl, drat)
}
