% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_vals.R
\name{get_id_vals}
\alias{get_id_vals}
\title{get_id_vals}
\usage{
get_id_vals(data, id)
}
\arguments{
\item{data}{dataframe}

\item{id}{id column}
}
\value{
dataframe of the render id
}
\description{
get_id_vals
}
\examples{
get_id_vals(df, render_id)
}
