% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_acc.R
\name{get_model_acc}
\alias{get_model_acc}
\title{get_model_acc}
\usage{
get_model_acc(df, label_var, pred_lab)
}
\arguments{
\item{df}{dataframe}

\item{label_var}{label variable}

\item{pred_lab}{the predicted label of model}
}
\value{
dataframe with model acc
}
\description{
get_model_acc
}
\examples{
get_model_acc(df, label_column, predicted_label)
}
