% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_cond.R
\name{get_model_cond}
\alias{get_model_cond}
\title{get_model_cond}
\usage{
get_model_cond(data, column_name)
}
\arguments{
\item{data}{dataframe}

\item{column_name}{column name for the conditions}
}
\value{
dataframe
}
\description{
get_model_cond
}
\examples{
get_model_cond(df, condition_column)
}
