% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_count.R
\name{get_model_count}
\alias{get_model_count}
\title{get_model_count}
\usage{
get_model_count(df, model_var)
}
\arguments{
\item{df}{dataframe}

\item{model_var}{column to count}
}
\value{
dataframe
}
\description{
get_model_count
}
\examples{
get_model_count(df, model_column)
}
