% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions_df.R
\name{get_predictions_df}
\alias{get_predictions_df}
\title{get_predictions_df}
\usage{
get_predictions_df(dataf, model_names)
}
\arguments{
\item{dataf}{dataframe}

\item{model_names}{name of which model}
}
\value{
dataframe of predicted values
}
\description{
get_predictions_df
}
\examples{
get_predictions_df(df, model)
}
