% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{rename_columns}
\usage{
rename_columns(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with renamed columns
}
\description{
rename_columns
}
\examples{
rename_columns(df)
}
