% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_labs.R
\name{rename_labs}
\alias{rename_labs}
\title{rename_labs}
\usage{
rename_labs(df, label)
}
\arguments{
\item{df}{dataframe}

\item{label}{label to rename}
}
\value{
renamed dataframe
}
\description{
rename_labs
}
\examples{
rename_labs(df, label)
}
