% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_questions.R
\name{rename_questions}
\alias{rename_questions}
\title{rename_questions}
\usage{
rename_questions(x, cond)
}
\arguments{
\item{x}{dataframe}

\item{cond}{condition}
}
\value{
renamed df
}
\description{
rename_questions
}
\examples{
rename_questions(df, 'FD1')
}
