% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_chi.R
\name{report_chi}
\alias{report_chi}
\title{report_chi}
\usage{
report_chi(test)
}
\arguments{
\item{test}{chi square test}
}
\value{
output of chi square test
}
\description{
report_chi
}
\examples{
chisq.test(c(A = 20, B = 15, C = 25)) |> report_chi()
}
