% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_conditions.R
\name{reverse_conditions}
\alias{reverse_conditions}
\title{reverse_conditions}
\usage{
reverse_conditions(data, cond_col, q1_val, q2_val)
}
\arguments{
\item{data}{dataframe}

\item{cond_col}{column of the condition}

\item{q1_val}{question 1 value}

\item{q2_val}{question 2 value}
}
\value{
dataframe with the counterbalanced conditions reversed
}
\description{
reverse_conditions
}
\examples{
reverse_conditions(df, condition_column, question1_value, question2_value)
}
