% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{show_progress}
\alias{show_progress}
\title{Determine progress bars should be shown}
\usage{
show_progress()
}
\description{
Progress bars are shown _unless_ one of the following is `TRUE`
- The bar is explicitly disabled by setting `options(vroom.show_progress = FALSE)`
- The code is run in a non-interactive session (`interactive()` is `FALSE`).
- The code is run in an RStudio notebook chunk.
- The code is run by knitr / rmarkdown.
- The code is run by testthat (the `TESTTHAT` envvar is `true`).
}
