% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_columns1.R
\name{subset_columns1}
\alias{subset_columns1}
\title{subset_columns1}
\usage{
subset_columns1(df, cond_ranges)
}
\arguments{
\item{df}{dataframe}

\item{cond_ranges}{data type that specifies the ranges for each condition. It is included in the package.#'}
}
\value{
df
}
\description{
subset_columns1
}
\examples{
subset_columns1(df, cond_ranges)
}
