% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_columns2.R
\name{subset_columns2}
\alias{subset_columns2}
\title{subset_columns2}
\usage{
subset_columns2(df, cond_ranges)
}
\arguments{
\item{df}{dataframe}

\item{cond_ranges}{data type that specifies the ranges for each condition. It is included in the package.}
}
\value{
df
}
\description{
subset_columns2
}
\examples{
subset_columns2(df, cond_ranges)
}
