% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_multi_wide.R
\name{survey_multi_wide}
\alias{survey_multi_wide}
\title{survey_multi_wide}
\usage{
survey_multi_wide(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
wide version of the input df
}
\description{
survey_multi_wide
}
\examples{
survey_multi_wide(df)
}
