% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_tidy.R
\name{t_test_tidy}
\alias{t_test_tidy}
\title{t_test_tidy}
\usage{
t_test_tidy(x)
}
\arguments{
\item{x}{t-test data type}
}
\value{
a tidy df for the t-test
}
\description{
t_test_tidy
}
\examples{
t_test_tidy(t.test(extra ~ group, data = sleep))
}
