% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_prop.R
\name{unique_prop}
\alias{unique_prop}
\title{unique_prop}
\usage{
unique_prop(x)
}
\arguments{
\item{x}{vector}
}
\value{
proportion of unique values in x
}
\description{
unique_prop
}
\examples{
mtcars[['carb']] |> unique_prop()
}
