% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom-package.R
\docType{package}
\name{vroom-package}
\alias{vroom-package}
\alias{_PACKAGE}
\title{vroom: Read and Write Rectangular Text Data Quickly}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The goal of 'vroom' is to read and write data (like
    'csv', 'tsv' and 'fwf') quickly. When reading it uses a quick initial
    indexing step, then reads the values lazily , so only the data you
    actually use needs to be read.  The writer formats the data in
    parallel and writes to disk asynchronously from formatting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vroom.r-lib.org}
  \item \url{https://github.com/r-lib/vroom}
  \item Report bugs at \url{https://github.com/r-lib/vroom/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item https://github.com/mandreyel/ (mio library) [copyright holder]
  \item R Core Team (localtime.c code adapted from R) [contributor]
  \item Jukka Jylänki (grisu3 implementation) [copyright holder]
  \item Mikkel Jørgensen (grisu3 implementation) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
