% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{vroom_example}
\alias{vroom_example}
\title{Get path to readr example}
\usage{
vroom_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If 'NULL', all available example files will be
listed.}
}
\description{
vroom comes bundled with a number of sample files in
its 'inst/extdata' directory. This function make them
easy to access
}
\examples{
vroom_example()
vroom_example("mtcars.csv")
}
