% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{vroom_str}
\alias{vroom_str}
\title{Structure of objects}
\usage{
vroom_str(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Similar to \code{str()} but with more information for Altrep objects.
}
\examples{
# when used on non-altrep objects altrep will always be false
vroom_str(mtcars)

mt <- vroom(vroom_example("mtcars.csv"), ",", altrep = c("chr", "dbl"))
vroom_str(mt)
}
