% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{guess_type}
\alias{guess_type}
\title{Guess the type of a vector}
\usage{
guess_type(x, na = c("", "NA"), locale = readr::default_locale(),
  guess_integer = FALSE)
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{guess_integer}{If \code{TRUE}, guess integer types for whole numbers, if
\code{FALSE} guess numeric type for all numbers.}
}
\description{
Guess the type of a vector
}
