% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom_write.R
\name{output_column}
\alias{output_column}
\title{Preprocess column for output}
\usage{
output_column(x)
}
\arguments{
\item{x}{A vector}
}
\description{
This is a generic function that applied to each column before it is saved
to disk. It provides a hook for S3 classes that need special handling.
}
\examples{
# Most types are returned unchanged
output_column(1)
output_column("x")

# datetimes are formatted in ISO 8601
output_column(Sys.Date())
output_column(Sys.time())
}
\keyword{internal}
