% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems}
\alias{problems}
\title{Retrieve parsing problems}
\usage{
problems(x = .Last.value, lazy = FALSE)
}
\arguments{
\item{x}{A data frame from \code{\link[=vroom]{vroom()}} or a character vector from
\code{\link[=vroom_lines]{vroom_lines()}}.}

\item{lazy}{If \code{TRUE}, just the problems found so far are returned. If
\code{FALSE} (the default) the lazy data is first read completely and all
problems are returned.}
}
\value{
A data frame with one row for each problem and six columns:
\itemize{
\item line - Physical line in the original input that caused the problem
\item row,col - Parsed row within the corresponding input and column that
caused the problem
\item expected - What vroom expected to find
\item actual - What it actually found
\item file - The file with the problem
}
}
\description{
vroom will only fail to parse a file if the file is invalid in a way that is
unrecoverable. However there are a number of non-fatal problems that you
might want to know about. You can retrieve a data frame of these problems
with this function.
}
