% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr.R, R/vroom_fwf.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{default_locale}
\alias{locale}
\alias{fwf_cols}
\alias{fwf_empty}
\alias{fwf_positions}
\alias{fwf_widths}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{readr}{\code{\link[readr]{default_locale}}, \code{\link[readr]{locale}}, \code{\link[readr]{fwf_cols}}, \code{\link[readr]{fwf_empty}}, \code{\link[readr]{fwf_positions}}, \code{\link[readr]{fwf_widths}}}
}}

