% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{vroom}
\alias{vroom}
\title{Read a tab delimited file into a tibbe}
\usage{
vroom(file, num_threads = parallel::detectCores())
}
\arguments{
\item{file}{path to a local file.}

\item{num_threads}{Number of threads to use when reading and materializing vectors.}
}
\description{
Read a tab delimited file into a tibbe
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}

readr::write_tsv(mtcars, "mtcars.tsv")
vroom("mtcars.tsv")

\dontshow{
unlink("mtcars.tsv")
setwd(.old_wd)
}
}
