% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{vroom_altrep_opts}
\alias{vroom_altrep_opts}
\title{Control which column types use Altrep}
\usage{
vroom_altrep_opts(which = NULL)
}
\arguments{
\item{which}{A character vector of column types to use Altrep for.}
}
\description{
Control which column types use Altrep
}
