% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{vroom_altrep_opts}
\alias{vroom_altrep_opts}
\title{Show which column types are using Altrep}
\usage{
vroom_altrep_opts(which = NULL)
}
\arguments{
\item{which}{A character vector of column types to use Altrep for. Can also
take \code{TRUE} or \code{FALSE} to use Altrep for all possible or none of the
types}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
This function is deprecated in favor of \code{vroom_altrep()}.
}
