% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom_fwf.R
\name{vroom_fwf}
\alias{vroom_fwf}
\alias{fwf_empty}
\alias{fwf_widths}
\alias{fwf_positions}
\alias{fwf_cols}
\title{Read a fixed-width file into a tibble}
\usage{
vroom_fwf(
  file,
  col_positions = fwf_empty(file, skip, n = guess_max),
  col_types = NULL,
  col_select = NULL,
  id = NULL,
  locale = default_locale(),
  na = c("", "NA"),
  comment = "",
  skip_empty_rows = TRUE,
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = 100,
  altrep = TRUE,
  num_threads = vroom_threads(),
  progress = vroom_progress(),
  show_col_types = NULL,
  .name_repair = "unique"
)

fwf_empty(file, skip = 0, col_names = NULL, comment = "", n = 100L)

fwf_widths(widths, col_names = NULL)

fwf_positions(start, end = NULL, col_names = NULL)

fwf_cols(...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a
single string or a raw vector). \code{file} can also be a character vector
containing multiple filepaths or a list containing multiple connections.

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will be automatically
decompressed. Files starting with \verb{http://}, \verb{https://}, \verb{ftp://}, or
\verb{ftps://} will be automatically downloaded. Remote compressed files
(\code{.gz}, \code{.bz2}, \code{.xz}, \code{.zip}) will be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. To be recognised as
literal data, wrap the input with \code{I()}.}

\item{col_positions}{Column positions, as created by \code{\link[=fwf_empty]{fwf_empty()}},
\code{fwf_widths()}, \code{fwf_positions()}, or \code{fwf_cols()}. To read in only
selected fields, use \code{fwf_positions()}. If the width of the last column
is variable (a ragged fwf file), supply the last end position as \code{NA}.}

\item{col_types}{One of \code{NULL}, a \code{\link[=cols]{cols()}} specification, or
a string.

If \code{NULL}, all column types will be inferred from \code{guess_max} rows
of the input, interspersed throughout the file. This is convenient (and
fast), but not robust. If the guessed types are wrong, you'll need to
increase \code{guess_max} or supply the correct types yourself.

Column specifications created by \code{\link[=list]{list()}} or \code{\link[=cols]{cols()}} must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[=cols_only]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
\itemize{
\item c = character
\item i = integer
\item I = big integer
\item n = number
\item d = double
\item l = logical
\item f = factor
\item D = date
\item T = date time
\item t = time
\item ? = guess
\item _ or - = skip
}

By default, reading a file without a column specification will print a
message showing the guessed types. To suppress this message, set
\code{show_col_types = FALSE}.}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}

\item{id}{Either a string or 'NULL'. If a string, the output will contain a
column with that name with the filename(s) as the value, i.e. this column
effectively tells you the source of each row. If 'NULL' (the default), no
such column will be created.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{comment}{A string used to identify comments. Any line that starts
with the comment string at the beginning of the file (before any data
lines) will be ignored. Unlike \code{\link[=vroom]{vroom()}}, comment lines in the middle
of the file are not filtered out.}

\item{skip_empty_rows}{Should blank rows be ignored altogether? i.e. If this
option is \code{TRUE} then blank rows will not be represented at all.  If it is
\code{FALSE} then they will be represented by \code{NA} values in all the columns.}

\item{trim_ws}{Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from
each field before parsing it?}

\item{skip}{Number of lines to skip before reading data. If \code{comment} is
supplied any commented lines are ignored \emph{after} skipping.}

\item{n_max}{Maximum number of lines to read.}

\item{guess_max}{Maximum number of lines to use for guessing column types.
See \code{vignette("column-types", package = "readr")} for more details.}

\item{altrep}{Control which column types use Altrep representations,
either a character vector of types, \code{TRUE} or \code{FALSE}. See
\code{\link[=vroom_altrep]{vroom_altrep()}} for for full details.}

\item{num_threads}{Number of threads to use when reading and materializing
vectors. If your data contains newlines within fields the parser will
automatically be forced to use a single thread only.}

\item{progress}{Display a progress bar? By default it will only display
in an interactive session and not while executing in an RStudio notebook
chunk. The display of the progress bar can be disabled by setting the
environment variable \code{VROOM_SHOW_PROGRESS} to \code{"false"}.}

\item{show_col_types}{Control showing the column specifications. If \code{TRUE}
column specifications are always shown, if \code{FALSE} they are never shown. If
\code{NULL} (the default), they are shown only if an explicit specification is
not given in \code{col_types}, i.e. if the types have been guessed.}

\item{.name_repair}{Handling of column names. The default behaviour is to
ensure column names are \code{"unique"}. Various repair strategies are
supported:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence of names.
\item \code{"unique"} (default value): Make sure names are unique and not empty.
\item \code{"check_unique"}: No name repair, but check they are \code{unique}.
\item \code{"unique_quiet"}: Repair with the \code{unique} strategy, quietly.
\item \code{"universal"}: Make the names \code{unique} and syntactic.
\item \code{"universal_quiet"}: Repair with the \code{universal} strategy, quietly.
\item A function: Apply custom name repair (e.g., \code{name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}.
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{col_names}{Either NULL, or a character vector column names.}

\item{n}{Number of lines the tokenizer will read to determine file structure. By default
it is set to 100.}

\item{widths}{Width of each field. Use \code{NA} as the width of the last field
when reading a ragged fixed-width file.}

\item{start, end}{Starting and ending (inclusive) positions of each field.
\strong{Positions are 1-based}: the first character in a line is at position 1.
Use \code{NA} as the last value of \code{end} when reading a ragged fixed-width
file.}

\item{...}{Named or unnamed arguments, each addressing one column. Each input
should be either a single integer (a column width) or a pair of integers
(column start and end positions). All arguments must have the same shape,
i.e. all widths or all positions.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}. Parsing problems are stored in the
\code{problems} attribute; use \code{\link[=problems]{problems()}} to inspect them.
}
\description{
Fixed-width files store tabular data with each field occupying a specific
range of character positions in every line. Once the fields are identified,
converting them to the appropriate R types works just like for delimited
files. The unique challenge with fixed-width files is describing where each
field begins and ends. \pkg{vroom} tries to ease this pain by offering a
few different ways to specify the field structure:
\itemize{
\item \code{fwf_empty()} - Guesses based on the positions of empty columns. This is
the default. (Note that \code{fwf_empty()} returns 0-based positions, for
internal use.)
\item \code{fwf_widths()} - Supply the widths of the columns.
\item \code{fwf_positions()} - Supply paired vectors of start and end positions. These
are interpreted as 1-based positions, so are off-by-one compared to the
output of \code{fwf_empty()}.
\item \code{fwf_cols()} - Supply named arguments of paired start and end positions or
column widths.
}

Note: \code{fwf_empty()} cannot work with a connection or with any of the input
types that involve a connection internally, which includes remote and
compressed files. The reason is that this would necessitate reading from the
connection twice. In these cases, you'll have to either provide the field
structure explicitly with another \verb{fwf_*()} function or download (and
decompress, if relevant) the file first.
}
\details{
Here's a enhanced example using the contents of the file accessed via
\code{vroom_example("fwf-sample.txt")}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{         1         2         3         4
123456789012345678901234567890123456789012
[     name 20      ][state 10][  ssn 12  ]
John Smith          WA        418-Y11-4111
Mary Hartford       CA        319-Z19-4341
Evan Nolan          IL        219-532-c301
}\if{html}{\out{</div>}}

Here are some valid field specifications for the above (they aren't all
equivalent! but they are all valid):

\if{html}{\out{<div class="sourceCode">}}\preformatted{fwf_widths(c(20, 10, 12), c("name", "state", "ssn"))
fwf_positions(c(1, 30), c(20, 42), c("name", "ssn"))
fwf_cols(state = c(21, 30), last = c(6, 20), first = c(1, 4), ssn = c(31, 42))
fwf_cols(name = c(1, 20), ssn = c(30, 42))
fwf_cols(name = 20, state = 10, ssn = 12)
}\if{html}{\out{</div>}}
}
\examples{
fwf_sample <- vroom_example("fwf-sample.txt")
writeLines(vroom_lines(fwf_sample))

# You can specify column positions in several ways:
# 1. Guess based on position of empty columns
vroom_fwf(fwf_sample, fwf_empty(fwf_sample, col_names = c("first", "last", "state", "ssn")))
# 2. A vector of field widths
vroom_fwf(fwf_sample, fwf_widths(c(20, 10, 12), c("name", "state", "ssn")))
# 3. Paired vectors of start and end positions
vroom_fwf(fwf_sample, fwf_positions(c(1, 30), c(20, 42), c("name", "ssn")))
# 4. Named arguments with start and end positions
vroom_fwf(fwf_sample, fwf_cols(name = c(1, 20), ssn = c(30, 42)))
# 5. Named arguments with column widths
vroom_fwf(fwf_sample, fwf_cols(name = 20, state = 10, ssn = 12))
}
