% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom_fwf.R
\name{vroom_fwf}
\alias{vroom_fwf}
\title{Read a fixed width file into a tibble}
\usage{
vroom_fwf(file, col_positions = fwf_empty(file[[1]], skip, n =
  guess_max), col_types = NULL, col_keep = NULL, col_skip = NULL,
  id = NULL, locale = default_locale(), na = c("", "NA"),
  comment = "", trim_ws = TRUE, skip = 0, n_max = Inf,
  guess_max = 100, num_threads = vroom_threads(),
  progress = vroom_progress(), .name_repair = "unique")
}
\arguments{
\item{file}{path to a local file.}

\item{col_types}{One of \code{NULL}, a \code{\link[=cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[=cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[=cols_only]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
c = character, i = integer, n = number, d = double,
l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or
\code{_}/\code{-} to skip the column.}

\item{col_keep}{Columns to keep in the output, all other columns will be
skipped. Input can be a character vector of column names, a logical vector
or a numeric vector of column indexes. Only one of `col_keep` or
`col_drop` can be used.}

\item{col_skip}{Columns to skip in the output, all other columns will be
kept. Input can be a character vector of column names, a logical vector
or a numeric vector of column indexes. Only one of `col_keep` or
`col_drop` can be used.}

\item{id}{Either a string or 'NULL'. If a string, the output will contain a
variable with that name with the filename(s) as the value. If 'NULL', the
default, no variable will be created.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}

\item{trim_ws}{Should leading and trailing whitespace be trimmed from
each field before parsing it?}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of records to read.}

\item{guess_max}{Maximum number of records to use for guessing column types.}

\item{num_threads}{Number of threads to use when reading and materializing vectors.}

\item{progress}{Display a progress bar? By default it will only display
in an interactive session and not while knitting a document. The display
is updated every 50,000 values and will only display if estimated reading
time is 5 seconds or more. The automatic progress bar can be disabled by
setting option \code{readr.show_progress} to \code{FALSE}.}

\item{.name_repair}{Handling of column names. By default, vroom ensures
column names are not empty and unique. See `.name_repair` as documented in
[tibble::tibble()] for additional options including supplying user defined
name repair functions.}
}
\description{
Read a fixed width file into a tibble
}
