% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{vroom_progress}
\alias{vroom_progress}
\title{Determine whether progress bars should be shown}
\usage{
vroom_progress()
}
\description{
By default, vroom shows progress bars. However, progress reporting is
suppressed if any of the following conditions hold:
\itemize{
\item The bar is explicitly disabled by setting the environment variable
\code{VROOM_SHOW_PROGRESS} to \code{"false"}.
\item The code is run in a non-interactive session, as determined by
\code{\link[rlang:is_interactive]{rlang::is_interactive()}}.
\item The code is run in an RStudio notebook chunk, as determined by
\code{getOption("rstudio.notebook.executing")}.
}
}
\examples{
vroom_progress()
}
