% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{vroom_progress}
\alias{vroom_progress}
\title{Determine if progress bars should be shown}
\usage{
vroom_progress()
}
\description{
Progress bars are shown \emph{unless} one of the following is \code{TRUE}
\itemize{
\item The bar is explicitly disabled by setting \code{Sys.getenv("VROOM_SHOW_PROGRESS"="false")}
\item The code is run in a non-interactive session (\code{interactive()} is \code{FALSE}).
\item The code is run in an RStudio notebook chunk.
\item The code is run by knitr / rmarkdown.
\item The code is run by testthat (the \code{TESTTHAT} envvar is \code{true}).
}
}
\examples{
vroom_progress()
}
