% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom_write.R
\name{vroom_write_lines}
\alias{vroom_write_lines}
\title{Write lines to a file}
\usage{
vroom_write_lines(x, file, eol = "\\n", na = "NA", append = FALSE)
}
\arguments{
\item{file}{path to a local file.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}
}
\description{
Write lines to a file
}
